\name{Travis_Dev-package}
\alias{Travis_Dev-package}
\alias{Travis_Dev}
\docType{package}
\title{
Travis_Dev
}
\description{
Transcriptomic View of Genomic Features
}
\details{
Transcriptomic View of Genomic Features
}
\author{
Jia Meng
Maintainer: Jia Meng <jia.meng@xjtlu.edu.cn> 
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
