\name{Travis_Dev-package}
\alias{Travis_Dev-package}
\alias{Travis_Dev}
\docType{package}
\title{
\packageTitle{Travis_Dev}
}
\description{
\packageDescription{Travis_Dev}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Travis_Dev}
\packageIndices{Travis_Dev}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Travis_Dev}

Maintainer: \packageMaintainer{Travis_Dev}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
